Make Pattern pretty-printing consistency tests JVM only
Should fix a build failure
