+++ tests/src/test/scala/regex/RegexPrettyPrinterTests.scala
@@ -2,21 +2,8 @@ package ceedubs.irrec
 package regex
 
 import Regex._
-import CharRegexGen._
 
 class RegexPrettyPrinterTests extends IrrecSuite {
-
-  test("char regex pretty printer matches java Pattern") {
-    forAll(genCharRegexAndMatch) { rm =>
-      val prettyRegex = rm.r.pprint
-      val regexHex = prettyRegex.map(_.toInt.toHexString).toList
-      val javaR = rm.r.toPattern
-      val candidateHex = rm.candidate.map(_.toInt.toHexString).toList
-      assert(
-        javaR.matcher(rm.candidate.mkString).matches,
-        s"${rm.candidate.mkString} ($candidateHex) should match $prettyRegex ($regexHex)")
-    }
-  }
 
   test("char regex pretty print examples") {
     lit('a').pprint should ===("a")
